First extremely rough cut of a QuickCheck demo.
This is an extremely rough cut at some sample code for a
(Haskell) QuickCheck demo given on May 20, 2010. Actually,
the code in ArithTests.hs is different in two ways:

1. The generator for Exps is now size-constrained, so it
   doesn't produce giant test data.

2. There are now some tests.

True to form, QuickCheck found that division by zero leads
to an exception. The next step would be to make division by
zero one of the possible results of evaluation, and to test
that it is signalled in the tests.